bool_refpts -> get_bool_refpts. Moved deprecated functions to deprecated.py
