+++ pyat/at/lattice/__init__.py
@@ -12,6 +12,7 @@ from .utils import *
 from .lattice_object import *
 from .cavity_access import *
 from .variable_elements import *
+from .deprecated import *
 # Define the module "lattice.constants" for backward compatibility
 # noinspection PyUnresolvedReferences
 from .. import constants
