+++ pyat/at/tracking/patpass.py
@@ -4,7 +4,7 @@ Simple parallelisation of atpass() using multiprocessing.
 from functools import partial
 import multiprocessing
 # noinspection PyProtectedMember
-from ..lattice.utils import _uint32_refs
+from ..lattice.utils import get_uint32_refpts
 from ..lattice import AtWarning, Element, DConstant, random
 from ..lattice import Refpts, End
 from warnings import warn
@@ -165,7 +165,7 @@ def patpass(lattice: Iterable[Element], r_in, nturns: int = 1,
 
     if not isinstance(lattice, list):
         lattice = list(lattice)
-    refpts = _uint32_refs(lattice, refpts)
+    refpts = get_uint32_refpts(lattice, refpts)
     bunch_currents = getattr(lattice, 'bunch_currents', np.zeros(1))
     bunch_spos = getattr(lattice, 'bunch_spos', np.zeros(1))
     kwargs.update(bunch_currents=bunch_currents, bunch_spos=bunch_spos)
